implemented ImportController and DeliciousImporter

+++ bibsonomy-webapp/src/main/java/org/bibsonomy/webapp/controller/actions/ImportController.java
@@ -12,27 +12,33 @@ import java.util.Map;
 import org.apache.commons.fileupload.FileItem;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.bibsonomy.common.enums.GroupingEntity;
 import org.bibsonomy.common.enums.PostUpdateOperation;
 import org.bibsonomy.importer.bookmark.file.FirefoxImporter;
 import org.bibsonomy.importer.bookmark.service.DeliciousImporterFactory;
 import org.bibsonomy.model.Bookmark;
 import org.bibsonomy.model.Post;
+import org.bibsonomy.model.Tag;
 import org.bibsonomy.model.logic.LogicInterface;
 import org.bibsonomy.services.importer.FileBookmarkImporter;
+import org.bibsonomy.services.importer.RelationImporter;
 import org.bibsonomy.services.importer.RemoteServiceBookmarkImporter;
 import org.bibsonomy.util.file.FileUploadInterface;
 import org.bibsonomy.util.file.HandleFileUpload;
 import org.bibsonomy.webapp.command.actions.ImportCommand;
+import org.bibsonomy.webapp.util.ErrorAware;
 import org.bibsonomy.webapp.util.MinimalisticController;
 import org.bibsonomy.webapp.util.View;
+import org.bibsonomy.webapp.validation.ImportValidator;
 import org.bibsonomy.webapp.view.Views;
+import org.springframework.validation.Errors;
 import org.springframework.web.multipart.commons.CommonsMultipartFile;
 
 /**
  * @author mwa
  * @version $Id$
  */
-public class ImportController implements MinimalisticController<ImportCommand>{
+public class ImportController implements MinimalisticController<ImportCommand>, ErrorAware{
 
 	private static final Log log = LogFactory.getLog(ImportController.class);
 
@@ -50,6 +56,8 @@ public class ImportController implements MinimalisticController<ImportCommand>{
 	 * path of the document
 	 */
 	private String docpath;
+		
+	private Errors errors = null;
 	
 	/**
 	 * Main method which starts the import
@@ -57,35 +65,81 @@ public class ImportController implements MinimalisticController<ImportCommand>{
 	public View workOn(ImportCommand command) {
 		
 		List<Post<Bookmark>> posts = new LinkedList<Post<Bookmark>>();
+		List<Tag> relations = new LinkedList<Tag>();
 		
-		if(command.getImportType().equals("delicious")){
-			try {
-				DeliciousImporterFactory importerFactory = new DeliciousImporterFactory();
-				RemoteServiceBookmarkImporter importer = importerFactory.getImporter();
-				importer.setCredentials(command.getUserName(), command.getPassWord());
-				posts = importer.getPosts();
-			} catch (IOException ex) {
-				log.error(ex.getMessage());
+		/** if importType is not set, the controller is loaded the first time
+		 * 	and no data is submitted **/
+		if(command.getImportType() != null){
+			
+			/** Validate the command **/
+			ImportValidator validator = new ImportValidator();
+			validator.validate(command, errors);
+					
+			/** display potential errors **/
+			if(errors.hasErrors()){
+				log.debug("Import couldn't get started due to existing errors in form");
+				errors.reject("error.requiredFields");
+				errors.rejectValue("errorMessage","error.requiredFields ");
+				return Views.IMPORT;
 			}
-		}else if(command.getImportType().equals("firefox")){
-			try{
-				File file = buildFile(command.getFile());
-				fileImporter = new FirefoxImporter();
-				fileImporter.initialize(file, command.getContext().getLoginUser(), command.getGrouping());
-				posts = fileImporter.getPosts();
+			
+			if(command.getImportType().equals("delicious")){
+				try {
+					DeliciousImporterFactory importerFactory = new DeliciousImporterFactory();
+					RemoteServiceBookmarkImporter importer = importerFactory.getImporter();
+					importer.setCredentials(command.getUserName(), command.getPassWord());
+					posts = importer.getPosts();
+					
+					if(command.isBundles()){
+						RelationImporter relationImporter = (RelationImporter) importerFactory.getImporter();
+						relations = relationImporter.getRelations();
+					}
+				} catch (IOException ex) {
+					errors.reject("error.furtherInformations");
+					log.error("IOEXception at ImportController: "+ex.getStackTrace());
+					command.setErrorMessage(ex.getMessage());
+				}
+			}else if(command.getImportType().equals("firefox")){
+				try{
+					File file = buildFile(command.getFile());
+					fileImporter = new FirefoxImporter();
+					fileImporter.initialize(file, command.getContext().getLoginUser(), command.getGrouping());
+					posts = fileImporter.getPosts();
 				
-				file.delete();
-			} catch (IOException ex) {
-				log.error(ex.getMessage());
+					file.delete();
+				} catch (IOException ex) {
+					errors.reject("error.furtherInformations");
+					command.setErrorMessage(ex.getMessage());
+				}
+			}
+		
+			/** how many posts where found? **/
+			command.setTotalCount(posts != null?posts.size():0);
+						
+			/** store the posts **/
+			storePosts(command, posts);	
+			
+			/** if available store relations **/
+			if(relations.size() > 0){
+				storeRelations(relations, command);
 			}
 		}
-				
-		if(!posts.isEmpty()){
-			storePosts(command, posts);
+		
+		return Views.IMPORT;	
+	}
+	
+	/**
+	 * Store the received bundles into the database
+	 * @param relations
+	 * @param command
+	 */
+	private void storeRelations(List<Tag> relations, ImportCommand command){
+		
+		command.setStoredConcepts(new LinkedList<String>());
+		for(Tag tag: relations){
+			String conceptName = this.logic.createConcept(tag, GroupingEntity.USER, command.getContext().getLoginUser().getName());
+			command.getStoredConcepts().add(conceptName);
 		}
-	
-		return Views.IMPORT_SUCCESS;
-	
 	}
 	
 	/**
@@ -106,21 +160,26 @@ public class ImportController implements MinimalisticController<ImportCommand>{
 		List<String> nonCreatedBookmarkEntries = new ArrayList<String>();
 		
 		for(Post<Bookmark> post: posts){
+			 
+			if(post.getUser() == null){
+				post.setUser(command.getContext().getLoginUser());
+			}
+			
 			List<?> singletonList = Collections.singletonList(post);
-			String bookmarkUrl = post.getResource().getUrl();
+			String title = post.getResource().getTitle();
 			try {
 				// throws an exception if the bookmark already exists in the
 				// system
 				List<String> createdPostHash = logic.createPosts((List<Post<?>>) singletonList);
-				newBookmarkEntries.put(createdPostHash.get(0), bookmarkUrl);
+				newBookmarkEntries.put(createdPostHash.get(0), title);
 			} catch (IllegalArgumentException e) {
 				// checks whether the update bookmarks checkbox is checked
 				if (command.isOverwrite()) {
 
 					List<String> createdPostHash = logic.updatePosts((List<Post<?>>) singletonList, PostUpdateOperation.UPDATE_ALL);
-					updatedBookmarkEntries.put(createdPostHash.get(0), bookmarkUrl);
+					updatedBookmarkEntries.put(createdPostHash.get(0), title);
 				} else {
-					nonCreatedBookmarkEntries.add(bookmarkUrl);
+					nonCreatedBookmarkEntries.add(title);
 				}
 			}
 		}
@@ -206,5 +265,16 @@ public class ImportController implements MinimalisticController<ImportCommand>{
 	public void setDocpath(String docpath) {
 		this.docpath = docpath;
 	}
+
+	@Override
+	public Errors getErrors() {
+		return errors;
+	}
+
+	@Override
+	public void setErrors(Errors errors) {
+		this.errors = errors;
+	}
+
 	
 }
